32bit last_msg_ms, gain tweaks, refactor check on good target data

+++ libraries/AC_PrecLand/AC_PrecLand_Companion.h
@@ -45,5 +45,5 @@ private:
     float               _distance_to_target;    // distance from the camera to target in meters
     uint64_t            _timestamp_us;          // timestamp when the image was captured(synced via UAVCAN)
     bool                _new_estimate;          // true if new data from the camera has been received
-    uint16_t            _last_msg_ms;           // the last time in ms an updated msg was received
+    uint32_t            _last_msg_ms;           // the last time in ms an updated msg was received
 };
